@@ -1,6 +1,6 @@
 {
   "name": "chatgpt-shortcut",
-  "version": "2.9.1",
+  "version": "3.0.0",
   "private": true,
   "scripts": {
     "docusaurus": "docusaurus",
@@ -11,37 +11,38 @@
     "clear": "docusaurus clear",
     "serve": "docusaurus serve",
     "write-translations": "docusaurus write-translations",
-    "write-heading-ids": "docusaurus write-heading-ids"
+    "write-heading-ids": "docusaurus write-heading-ids",
+    "dev": "docusaurus start"
   },
   "dependencies": {
     "@ant-design/compatible": "^5.1.2",
-    "@docusaurus/core": "^2.4.3",
-    "@docusaurus/plugin-client-redirects": "^2.4.3",
-    "@docusaurus/preset-classic": "^2.4.3",
+    "@docusaurus/core": "^3.0.0",
+    "@docusaurus/plugin-client-redirects": "^3.0.0",
+    "@docusaurus/preset-classic": "^3.0.0",
     "@emoji-mart/data": "^1.1.2",
     "@emoji-mart/react": "^1.1.1",
-    "@mdx-js/react": "^1.6.22",
+    "@mdx-js/react": "^3.0.0",
     "@popperjs/core": "^2.11.6",
-    "@waline/vercel": "^1.31.3",
-    "antd": "^5.9.4",
-    "axios": "^1.5.1",
+    "@waline/vercel": "^1.31.7",
+    "antd": "^5.10.3",
+    "axios": "^1.6.0",
     "clsx": "^1.2.1",
-    "docusaurus2-plugin-google-adsense": "^1.1.0",
     "emoji-mart": "^5.5.2",
     "js-cookie": "^3.0.5",
     "lodash": "^4.0.0",
     "marked": "^5.1.2",
-    "prism-react-renderer": "^1.3.5",
-    "react": "^17.0.2",
+    "prism-react-renderer": "^2.1.0",
+    "react": "^18.2.0",
     "react-copy-to-clipboard": "^5.1.0",
-    "react-dom": "^17.0.2",
+    "react-dom": "^18.2.0",
     "react-giphy-searchbox": "^1.5.4",
     "react-html-parser": "^2.0.2",
     "react-popper": "^2.2.5",
     "react-share": "^4.4.1"
   },
   "devDependencies": {
-    "@docusaurus/module-type-aliases": "^2.4.3",
+    "@docusaurus/module-type-aliases": "^3.0.0",
+    "@docusaurus/types": "3.0.0",
     "@waline/client": "^2.15.7"
   },
   "browserslist": {
@@ -51,12 +52,12 @@
       "not op_mini all"
     ],
     "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
+      "last 3 chrome version",
+      "last 3 firefox version",
+      "last 5 safari version"
     ]
   },
   "engines": {
-    "node": ">=16.14"
+    "node": ">=18.0"
   }
 }
