@@ -170,7 +170,7 @@ function useSiteCountPlural() {
         {
           id: "showcase.filters.resultCount",
           description:
-            'Pluralized label for the number of sites found on the showcase. Use as much plural forms (separated by "|") as your language support (see https://www.unicode.org/cldr/cldr-aux/charts/34/supplemental/language_plural_rules.html)',
+            'Pluralized label for the number of prompts found on the showcase. Use as much plural forms (separated by "|") as your language support (see https://www.unicode.org/cldr/cldr-aux/charts/34/supplemental/language_plural_rules.html)',
           message: "{sitesCount} prompts",
         },
         { sitesCount }
@@ -214,7 +214,7 @@ function ShowcaseFilters({ onToggleDescription }) {
             className={styles.onToggleButton}
             title={translate({
               id: "toggle_prompt_language_description",
-              message: "更改提示词的显示语言，在英语和页面当前语言之间切换。",
+              message: "更改提示词的显示语言，可以在英语和当前页面语言之间进行切换。",
             })}>
             {togglePromptLanguage}
           </button>
@@ -492,7 +492,7 @@ function ShowcaseCards({ isDescription }) {
                 )}>
                 <Heading as='h2'>
                   <Translate id='showcase.favoritesList.title'>
-                    Our favorites
+                    Favorites
                   </Translate>
                 </Heading>
                 <FavoriteIcon svgClass={styles.svgIconFavorite} />
